restoring Go lab 10 to mimic Python lab
Python lab main.py file provided initial example code for set_announcement (including import, method declaration, and init method to add it as a handler function in the http package. Go did not. I am restoring that model to Go. This is the first step of 3 code blocks to move.